Move cache directly under mantis dir
