+++ src/mantis/jira/utils/__init__.py
@@ -1,2 +1,1 @@
-from .cache import Cache
 from .jira_system_config_loader import JiraSystemConfigLoader
