+++ tests/test_jira_inspector.py
@@ -4,7 +4,7 @@ import pytest
 from typing import Any
 
 from mantis.jira import JiraClient
-from mantis.jira.utils.cache import CacheMissException
+from mantis.cache import CacheMissException
 from mantis.jira.utils.jira_system_config_loader import Inspector
 from tests.data import CacheData
